Better and doc for standard option factories
